adjusting airlines_in_2020 functions to seperate file
